feat(read-ahead)(2/N): Infer localfile sequential read to activate read ahead (#456)
# Introduction

In #411, a read-ahead mechanism was introduced in the localfile layer to improve RPC latency for sequential reads.
This PR further enhances that by enabling the system to recognize sequential read patterns and apply read-ahead for the identified applications.

The `spark.rss.riffle.readAheadEnabled` (it was disabled by default) parameter is introduced for the Uniffle client.
It controls whether the read-ahead mechanism is enabled on the Riffle server, providing a simple way for the client to toggle this feature.

# Commits

* feat(read-ahead)(2/N): Infer localfile sequential read from memory read (followup: https://github.com/zuston/riffle/pull/411)

* refactor read options

* create read options new file

* activate it

* introduce dedicate app level option to control whether read ahead is enabled

* add app level test